+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/AzureMonitorOptions.cs
@@ -61,16 +61,15 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore
         /// <summary>
         /// Gets or sets the ratio of telemetry items to be sampled. The value must be between 0.0F and 1.0F, inclusive.
         /// For example, specifying 0.4 means that 40% of traces are sampled and 60% are dropped.
-        /// The default value is 1.0F, indicating that all telemetry items are sampled.
         /// </summary>
-        public float SamplingRatio { get; set; } = 1.0F;
+        public float? SamplingRatio { get; set; }
 
         /// <summary>
         /// Gets or sets the number of traces per second to be sampled when using rate-limited sampling.
         /// For example, specifying 0.5 means one request every two seconds.
         /// When both TracesPerSecond and SamplingRatio are specified, TracesPerSecond takes precedence.
         /// </summary>
-        public double? TracesPerSecond { get; set; } = 5.0;
+        public double? TracesPerSecond { get; set; }
 
         /// <summary>
         /// Override the default directory for offline storage.